adding error handler

+++ backend/routers/post.js
@@ -32,7 +32,8 @@ router.post('', CHECK_AUTH, multer({storage: storage}).single('image'), (req, re
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
     res.status(201).json({
@@ -41,6 +42,11 @@ router.post('', CHECK_AUTH, multer({storage: storage}).single('image'), (req, re
         ...createdPost,
         id: createdPost._id
       }
+    });
+  })
+  .catch(err => {
+    res.status(500).json({
+      message: 'Creating a post failed!'
     });
   });
 });
@@ -67,6 +73,11 @@ router.get('',(req, res, next) => {
         count: count
       });
     })
+    .catch(err => {
+      res.status(500).json({
+        message: 'Fetching posts failed!'
+      });
+    });
 });
 
 router.put('/:id', CHECK_AUTH, multer({storage: storage}).single('image'), (req, res, next) => {
@@ -79,11 +90,23 @@ router.put('/:id', CHECK_AUTH, multer({storage: storage}).single('image'), (req,
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({
-      message: 'Post updated successfully'
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: 'Post updated successfully'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized!'
+      });
+    }
+  })
+  .catch(err => {
+    res.status(500).json({
+      message: 'Updating a post failed!'
     });
   });
 });
@@ -96,12 +119,26 @@ router.get('/:id', (req, res, next) => {
       res.status(404).json({message: 'Post not found'});
     }
   })
+  .catch(err => {
+    res.status(500).json({
+      message: 'Fetching post faild!'
+    });
+  });
 });
 
 router.delete('/:id', CHECK_AUTH, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({message: 'Post deleted!'});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Post deleted!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   })
+  .catch(err => {
+    res.status(500).json({
+      message: 'Deleting a post failed!'
+    });
+  });
 });
 
 module.exports = router;
